examples/strands-agent/requirements.txt
strands-agents>=1.0.0
openai>=1.40.0
